schematics: make autocrafter stop faster
I now really stops at the row with missing items
by using an instant NOR gate (pistons cutting
redstone dust) and removing the block powering the
rails when the auto stop triggers (instead of just
stopping the clock), yet this is still 2 ticks
short of stopping it before the row with missing
items.

Also, the UI is more compact.